addressed comments on the first Form Layout PR
