+++ backend/src/types/enums.ts
@@ -1,0 +1,18 @@
+export enum EventStatus {
+  UPCOMING = "UPCOMING",
+  ACTIVE = "ACTIVE",
+  CLOSED = "CLOSED",
+  CANCELLED = "CANCELLED",
+}
+
+export enum FormStatus {
+  OPEN = "OPEN",
+  CLOSED = "CLOSED",
+}
+
+export enum QuestionType {
+  SHORT_TEXT = "SHORT_TEXT",
+  LONG_TEXT = "LONG_TEXT",
+  EMAIL = "EMAIL",
+  DROP_DOWN = "DROP_DOWN",
+}
