--- backend/src/types/eventStatus.ts
@@ -1,6 +1,0 @@
-export enum EventStatus {
-  UPCOMING = "UPCOMING",
-  ACTIVE = "ACTIVE",
-  CLOSED = "CLOSED",
-  CANCELLED = "CANCELLED",
-}
