--- backend/src/types/formStatus.ts
@@ -1,4 +1,0 @@
-export enum FormStatus {
-  OPEN = "OPEN",
-  CLOSED = "CLOSED",
-}
