--- backend/src/types/questionType.ts
@@ -1,6 +1,0 @@
-export enum QuestionType {
-  SHORT_TEXT = "SHORT_TEXT",
-  LONG_TEXT = "LONG_TEXT",
-  EMAIL = "EMAIL",
-  DROP_DOWN = "DROP_DOWN",
-}
