+++ backend/src/utils/resMessages.ts
@@ -1,11 +1,7 @@
 export const errorMessage = (message: string) => {
-  return { message: message}
+  return { message: message }
 }
 
 export const serverErrorMessage = (err: any) => {
   return { message: err.message }
-}
-
-export const successMessage = (message: string) => {
-  return { message: message } 
 }
